model: VirtualFileHistory(Node|Edge) moved to separate files
This considerably reduces the size of the class VirtualFileHistoryGraph.
